Add ThreadPoolTaskScheduler implementing TaskScheduler class.
ThreadPoolTaskScheduler is an async task handler with user-defined amount of
concurrent threads listening to a SyncQueue that is fed by the user with
tasks to be executed trough the TaskScheduler API. Each thread started gets a distinct name,
this way it is easier to profile or debug the thread pool in case of an error or in a
complex system to track which component uses to much CPU time. Thread name setup is currently
only supported for Linux, QNX and macOS.
Two basic unit tests scenarios with single-user push and multi-user push also added.

Relates-to: OLPEDGE-394

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>