Rename NetworkStatus to SendOutcome to mark it is a outcome class.
NetworkStatus implied that the return indicates if Network is online/offline
and not that the Send() request was successfull or failed. Renamed to SendOutcome
to mark that it is a outcome pattern applied on to the Send() method.
This also adapts doxygen comments and certain variable names to make
them more explicit.

Relates-to: OLPEDGE-369

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>