Fixes the cache clear key for partitions and the according test.
When adding a layers partition metadata response to the cache the
key is hrn::layer::version::partitions. In case we receive a 403 error
we try to delete the entire entry from the cache with RemoveKeysWithPrefix()
were the key was hrn::layer::::partitions which didn't work as the entries
include version also. Changing the key in case no version is provided to
hrn::layer:: which will get successfull prefix comparison in the cache impl.

Relates-to: OLPEDGE-643

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>