VersionedLayerClient class implementation
The class' GetData method is introduced.
It fetches data from a layer using a partition id or data handle.
The data is fetched asynchronously if the task scheduler is set
properly. The synchronous implementation of the logic is placed to
DataRepository::GetData() as it fits better from an architectural
point of view.

Resolves: OLPEDGE-759

Signed-off-by: Sergii Vostrikov <ext-sergii.vostrikov@here.com>