Adapts StreamLayerClient to use TaskScheduler for Flush context.
This removed the raw std::thread usage and triggeres instead a task
to be executed in the context of the TaskScheduler. AutoFlushController
is the only class remaining that still uses raw std::thread but this
will need a big refactoring and will be postponed post v0.7.0 release.
Additionaly this adds a multi-publish usecase with flush to the write
example.

Relates-to: OLPEDGE-457

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>