Adapts InMemoryCache class to coding style.
Adapts most of the internals of the InMemoryCache to the coding style.
Also removes static methods from InMemoryCache and replaces them with
nested structures to be used by the LRUCache and time provider default
parameter values. Add mutex lock for Size() method.
Promote private members to protected to be able to test via testable.
Adapts unit tests to coding style.

Relates-to: OLPEDGE-566

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>