Small changes needed for VersionedLayerClient::PrefetchTiles.
Allows injecting CancellationContext into TaskContext so that we can
share a single CancellationContext for multiple tasks. This is mandatory
for the PrefetchTiles where one CancellationToken, resp. CancellationContext,
is handed over to the caller and should be able to cancelled all scheduled
tasks.
Changes ApiResponse from const reference on the ResultType to by copy passing
so that we can move effectively OLP responses.
Modifies CancellationContext::ExecuteOrCancelled() to return bool denoting
if operation was successfully executed or cancelled.

Relates-to: OLPEDGE-1005

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>