Refactors VersionedLayerClient::PrefetchTiles now using TaskScheduler.
Instead of using std::async which creates countless threads that are not
bound to anything and still perform even when client is destroyed.
This commit tries to use the provided TaskScheduler to perform the requests.
This is done all in an initial task to gather all necessary information like
latest catalog version if not set in the request and the quadkey metadata after
all provided tiles are merged as much as possible to common ancestors. Then for
each quadkey a own task is scheduled with an additional last task that waits on
all quadkeys to be retrieved and provides the overall result to the user via the
provided callback.

- Add new sync method GetSubTiles() to PrefetchTilesRepository.
- Change PrefetchTilesProvider to use raw CancellationContext.
- Remove PrefetchTilesBusy test as current implementation can be
  triggered mutiple times without the need to wait for previous to finish.
- Added AddTask() template to VersionedLayerClientImpl to schedule task.

Relates-to: OLPEDGE-1005

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>