Refactored assertion about GetBatch successful response in tests
Refactored the assertion about waiting for the successful result of GetBatch response in functional tests.
Now we only check whether the GetBatch response from serve was either 'submitted' or 'succeeded'.
We can't rely on the continuous waiting for the 'succeeded' state as it might not be finished in a determined way.

Resolves: OLPEDGE-708

Signed-off-by: Bohdan Kurylovych <ext-bohdan.kurylovych@here.com>