Add StreamLayerClient and StreamLayerClientImpl class skeletons
StreamLayerClient provides the ability to consume data from a stream layer
in real time. The client reads the data in the order it is added to the
queue. Once it reads the data, the data is no longer available to
that client, but the data remains available to other clients.

The StreamLayerClient is movable-only since the each client is distinguished by
the back-end by its personal ID (consumerID) which is required to be unique within
the same group (the consumer group).

Resolves: OLPEDGE-1349

Signed-off-by: Dmytro Poberezhnyi <ext-dmytro.poberezhnyi@here.com>