Adapts HRN, CancellationContext and Condition to coding style.
This commit adapt the above mentioned classes the the coding style,
also marks the public members of HRN as deprecated and to be
scoped to private by 05.2020 as it is not recommended by HERE Platform
to interpret the HRN in any way. Users should only use the public methods
available and not care about the composition of the HRN string.

Relates-to: OLPEDGE-567

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>