Changes DefaultCache constructor to take settings by value.
Additionally makes constructor explicit. It is not a good practice
to take hide move oportunities from API users. Making the DefaultCache
constructor and the OlpClientSettingsFactory::CreateDefaultCache() method
take-in the CacheSettings by copy will allow user to pass-move and save
possible copies on the strings.
This also deprecates the olp::dataservice::write::StreamLayerClient
CreateDefaultCache() function as it is a duplicate to the
OlpClientSettingsFactory::CreateDefaultCache() method. Will be removed
by 06.2020. Until then this free function will use OlpClientSettingsFactory
internally to create the default cache instance.

Resolves: OLPEDGE-1509

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>
Signed-off-by: Andrei Popescu <andrei.popescu@here.com>