Deprecate WithVersion methods in request classes.
VersionedLayerClient needs to serve consistent data for all requests
which are triggered on a particular instance. If user do not know the
catalog version up front and always use boost::none for version, the
risk of returning different catalog versions data is high. Therefore
we lock the catalog version in the constructor of the VersionLayerClient
instance and all requests are therefore served with that particular catalog
version. Hence we do not need the WithVersion() methods in the requests
structures anymore and they will be ignore during requests and the constructor
passed catalog version will be taken instead.

Relates-To: OLPEDGE-1514

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>
Signed-off-by: Andrei Popescu <andrei.popescu@here.com>