Remove trailing whitespaces.
Trailing whitespaces are evil.
Don't commit evil into your repo.

Trailing whitespace issues can cause a lot of problems when they get
into your repository. It leads to false diffs which claim lines have
been changed when in fact the only the thing that changed was spacing.

This can make finding what actually changed in a file later on in
the development cycle next to impossible.

Relates-To: OLPEDGE-1595

Signed-off-by: Kirill Zhuchkov <kirill.zhuchkov@here.com>