Add support for a list of partitions to GetPartition API
When the user specifies the list of partitions the query API is used instead of
metadata API. This brings the user a functionality to access partitions
metadata of individual partitions.
The implementation checks the cache, if the number of requested partitions is
less then the number of partitions found in cache, the check fails and SDK
performs the online request.

Resolves: OLPEDGE-1017

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>