Adapts TokenEndpoint to coding style.
TokenEndpoint impl is now a class, not a struct anymore as this is
against the coding style. Some minor variables renamed. This also
moves the impl methods definition from inlined to ouside the class
to align all methods.
TokenEndpoint, as it is currently public, should also include its
dependency headers via global scope.

Relates-to: OLPEDGE-1028

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>