Adapts OlpClient to the coding style and simplifies internals. (#683)
Two unnecessary moves in ApiClientLookup::LookupApi() removed, as this resulted
together with changing OlpClient::SetSettings() take settings in as copy to crash
as this will move twice the same resource which is not possible.
Added aliases for input parameters on OlpClient::CallApi() to make the footprint
smaller. Internally refactored a little to make better use of the method chaining
from NetworkSettings and NetworkRequest. Moved Bearer header to a free function.
Move constructor and destructor default definition to cpp file.

Relates-to: OLPEDGE-567

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>