Reduce compiler warnings about deprecated methods/classes.
As we deprecated some classes that are public but only to be used
internaly there are a lot of deprecated warnings when building the SDK.
This adds pragma compiler flag disables for the places where we actually
use these classes internaly until we can finaly move these classes to private.
Also this commit removes the version usage from the request key creation and
adds again pragma disables to the repositories which are still using the version
setters/getters internaly. This will be changed in the future but until then
we need to lower the warnings count while building.

Resolves: OLPSUP-9858

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>