Apply pimpl patter to TokenProvider.
To be able to share one single instance of TokenProvider with multiple
users a shared pimpl is applied. This is especially helpfull for
multiple layer clients that access the same catalog with the same
credentials and in turn the same token. This will save a huge amount
of token requests.

Relates-To: OLPEDGE-1791

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>