Reduce excesive logging in API repositories. (#848)
The SDK logs to much and to often especially in the API repositories.
This commit removes some of the duplicated logs, aligns all logs to a
common structure and adds additional info like HRN and key to all
important logs. All other logs that are not important are now on level
debug. All previous trace logs which are actually printing debug level
information are now raised to debug level.

One very important fix that needs highlight is the capture by copy
which was removed from most of the cache repositories while putting
into cache. It is not necessary to capture model data by copy as
the KeyValueCache::Put() is pure sync and will not leave the context
so that the model faces deletion. This additional copy is not needed
and might even bring performance decreases. This is now fixed and all
models are captured by reference when added to cache.

Additionally this commit also enables ccplint namespace check.

Resolves: OLPSUP-10501

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>