Add compression settings to CacheSettings.
This commit adds the possible compression settings to the CacheSettings
structure to be passed on to leveldb. Possible values so far are only
no compression or Snappy compression. Nevertheless there are user who
will profit from not applying any compression if the data added to the
cache is already compressed, e.g. protobuf.

Additionally this commit adds filter policy to bloom filter with 10
bits precision which will enhance reading speeds.

Relates-To: OLPEDGE-1984, OLPEDGE-1985

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>