Optimize reading data from leveldb with iterator.
When reading data from leveldb::Get() we need to use a string which
is poorly initialized and data is copied slowly and inefficient to it
which then again is memcpy-ied to a shared vector which adds another
layer of memcpy and slow initialization. The overhead for frequent data
reads is more then enough to consider an alternative.
This alternative consist in reading the data directly from the DB into
the resulting shared vector without any intermediate steps by using
iterators. The leveldb iterator gives us the possibility to access the
raw DB data and size without passing it into a inefficient string
before, and gives a huge performance bump.

Relates-To: OLPEDGE-1983

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>