Remove expired elements during cache eviction (#870)
Remove expired elements during cache eviction

When eviction happens, first we remove all expired elements, and only
then we remove other elements if still needed.
Change the internal LRU representation, so that we have fast access to
key expiration. Saves extra cache lookup during eviction.
Fix the use case when protected cache contains expired data.

Adds compact API and automatic compacting.
Compacting the database is important in case a lot of operations are
performed in a small time frame, increasing the database size and
making the internal indexes unaligned. Doing this on a regular basis
without waiting for leveldb internal runners to perform this at a
unknown time we can do it on our own before a batch is applied in a
separate thread and when the database is closed. This way once the
database is opened next time the indexes are rebalanced and fast to
access.
Also, this commits exposes the compact API to the DefaultCache users
in case they want to do it themselves on certain checkpoints in time.

Relates-To: OLPEDGE-1925

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>
Co-authored-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>