Change to using system time by default for oauth.
Until now the default behaviour for oauth trough TokenProvider or
AuthenticationClient was to use the server timestamp and only on user
request or on error use the system time as an alternative.
This behaviour is now reversed so that as a default system time is
used and users need to explicitly change the use_system_time variable
in AuthenticationSettings or TokenProvider Settings to false to use
server time in case they know that the system time is not to be
trusted.

Relates-To: OLPEDGE-2009

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>