Make OlpClient a shared pimpl class.
In the scope of merging same requests, to avoid multiple downloads of
same data, OlpClient with its URL based API and used by all repos
for request will play a major role in merging this requests, i.e.
keeping an own list of pending requests. For this a more effective
way of sharing it is needed, except wrapping it in std::shared_ptr.
Thus turning OlpClient into a shared pimpl class is an easy way to
achievethe desired level of shareability.

Relates-To: OLPEDGE-1805

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>