Add Protect/Release functionality (#958)
Add interface functions Protect/Release.
Protect used mark keys as non-evictable that they should stay in cache,
till their state will be changed back after calling Release with that 
keys.

Calling Protect function with some keys and prefixes
adds them to protected list, which guarantee that keys
will never be evicted or expired, even if LRU eviction is used, and 
expiration is set. Keys could be protected before, or after they were
added to cache. If key was protected before, and expiration is set, 
it will be written to cache with expiration, but never removed from 
cache after time elapsed. If key was protected after, and still in 
cache, after expiration it will be returned to user, without 
considering expiration. In both cases protected keys are not added 
to lru, so they will never be evicted.

With Release function protected keys will be removed from
protected list, which means they could be evicted by lru,
and will expire after, if expiration set.

Add integration tests to check if protected keys
are not expire and are not evictable.

Relates-To: OLPEDGE-2124

Signed-off-by: Liubov Didkivska ext-liubov.didkivska@here.com