Add api to check if partition/tile cached (#967)
Add functionality to check if certain tile/partition is cached.

For partition we are checking if data handle is found in cache,
if data handle exist, check if data for corresponding partition
in cache.

For tile key search corresponding QuadTreeIndex,
considering that depth for all cached QuadTreeIndex is 4.
If quad tree is found, which should contain our tile, read
data handle for that tile. If data handle exist, check if data
for corresponding tile is cached.

Add integration tests for to check if partition and tile is cached.
Get partition and tile. Check if it is in cache using IsCached
function. Remove partition/tile with RemoveFromCache function and
check again that it is no longer cached.

Relates-To: OLPEDGE-2111

Signed-off-by: Liubov Didkivska ext-liubov.didkivska@here.com