Change Prefetch default levels (#980)
Prefetch works with input parameters min
and max levels. Defaults levels should
indicate that user wants to get only
tiles in request. As level 0 is valid,
min and max values for request is
initialized as invalid, that during
reques we can distinguish if levels
was set by user.
Fix tests to run with default initialized
levels instead of 0/0.

Relates-To: OLPEDGE-2198

Signed-off-by: Liubov Didkivska <ext-liubov.didkivska@here.com>