Add API to Protect/Release tile keys
Protect adds tile keys to protected list.
If some tile key was marked as protected, it is expected to stay in
cache, that its data can be accessible.
Release keys, which was marked before as protected. After releasing tile
keys associated data in cache could expire or be evicted.

Protect and Release functionality searches for quad tree in cache to get
data handle for each tile. When protecting tiles, at least one tile is
needed to be protected to also add quad to protected list. When
releasing all tiles should be not protected to release quad.

Added integration test with protecting and releasing tiles
keys under common quad tree

Relates-To: OLPEDGE-2127

Signed-off-by: Liubov Didkivska <ext-liubov.didkivska@here.com>