Adding CatalogEndpointProvider.
Api url can be static. In such cases we can avoid lookup request and
save some time. For that purpose we are adding CatalogEndpointProvider
to ApiLookupSettings.

ApiLookupClient will try to get url using the provider from settings
instead of making api lookup request. If it can't be done, behavior
of the client is unchanged.

Also adding unit and integration test for ApiLookupClient.

Resolves: OLPEDGE-2187

Signed-off-by: Kostiantyn Zvieriev <ext-kostiantyn.zvieriev@here.com>