Add Content-Encoding header to StreamLayerClient (#1003)
Automatically add Content-Encoding header to StreamLayerClient,
needed for gzip streams (when the Stream layer is configured as
compressed on the platform). Without this header the server responds
with status 400 Bad Request. Currently user have to manually compress
the payload that he wants to send, however when downloading, the
content is unziped automatically by network libraries.

Resolves: OLPEDGE-11219

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>