Correct narraow cast from uint64_t to size_t.
On 32Bit targets size_t is 32 bits, resp. unsigned int which results
in a -Wc++11-narrowing error when it gets feed with uint64_t values.
This commits safly narrow casts by trimming to std::numeric_limits
max for size_t which in case of 64Bit platforms will result in no
data loss while on 32Bit it gets trimmed by 32 upper bits.
This is ok in this case as the value is a counter for downloaded
bytes which is still more then 4GB. We do not expect such huge
values to be downloaded in one shot either way.

Relates-To: OLPEDGE-2169

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>