Fix the IsCached behavior for aggregated tiles
Fix the IsCached behavior for aggregated tiles when the aggregated
tile root is too far away, for example the tile on level 14 is
aggregated in tile on level 1, it means that the tile could be found
on levels 10-14, and tiles on levels 1-9 can't be found anymore.
In such case we iterate up and fetch quad trees until the aggregated
tile root is found in tree subquads. So it could be found later by
other APIs, ex. Protect, Release, Remove, etc.

Resolves: OLPSUP-11651

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>