Inverse the GetLatestVersion logic.
Change the logic of GetLatestVersion, so in the case of fetch option is
OnlineIfNotFound, we check the online first, then in case of an error,
the cache is checked. It is needed in the use case when you need to get
the latest data from online, and be able to use the data from cache.

Remove the five minutes expiration of the latest version, this enables
the users to work with data offline with ease.

Relates-To: OLPEDGE-2325, OLPEDGE-2326

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>