DefaultCache Open()/Close() individual caches.
Adding new Open/Close API for DefaultCache. Now it's possible to
disable protected or mutable cache, modify and enable it again. This
can be useful if for some reason we want to update individual cache
while DefaultCache is still in use.

Changed DiskCache behavior when try to open not existing folder for
read-only. Now it's creating folder and an empty leveldb DB.

Removed CreadteDir() function, since it's not used anymore and also
we have alternative util for that purpose(olp::utils::Dir::create()).

Resolves: OAM-731, OLPSUP-12312

Signed-off-by: Kostiantyn Zvieriev <ext-kostiantyn.zvieriev@here.com>