Enable OlpClient caching ApiLookupClient
This enables SDK to route user request to the same OlpClient instances,
that is needed to implement a proper requests merging. Another benefit
for users is less cache lookups. The client is cached for max-age time
provided by the server, or one hour. When the URL is resolved from the
cache we assume it will expire in five minutes.
Move DefaultLookupEndpointProvider source code to cpp.

Resolves: OLPEDGE-2230

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>