Enable OlpClient caching ApiLookupClient (#1107)
This enables SDK to route user requests to the same OlpClient instances.
Needed to implement a proper request merging. Another benefit
for users is fewer cache lookups. The client is cached for a max-age time
provided by the server or one hour. When the URL is resolved from the
cache, we assume it will expire in five minutes.
Move DefaultLookupEndpointProvider source code to the cpp file.

Resolves: OLPEDGE-2230

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>