Fix NetworkMock issue resulting in request lock.
This commit fixes a wrong usage of the NetworkMock::ReturnHttpResponse
which does not have any sleep time by default and resulted in optimal
cases in callback trigger prior to the Network::Send mock call
returning which caught the OlpClient PendingUrlRequest with no valid
request id set and thus resulted in a lock in waiting for the callback
which never came.

Some other tests need additional increasing of wait timeout as the
delay introduced in NetworkMock response increases the overall
execution.

Additionaly this adds some more debug logs to PendingUrlRequests class
to see what callback is added or cancelled and changes the
ApiLookupClientTest.LookupApiAsync to really use the async API as this
did not happen due to copy-paste code from LookupApi test.

Resolves: OLPEDGE-2409

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>