Add methods to get/set size of the mutable cache (#1129)
Using Size(CacheType) method users will be able to check current size
of the data stored in the mutable cache. Note that the returned size
is approximate and may not match cache folder size. It is the size
of real data and does not include levelDB overhead.

Using Size(uint64_t) method users will be able to change maximum
alowed size of the mutable cache on a fly. If size increases, no side
actions happen. If size decreases, the data that doesn't fit into new
size (taking into account data eviction thresholds) will be exicted
from the cache. If a lot of data must be evicted, it will be splitted
into chunks that fits into CacheSettings::max_eviction_portion.

Note that implementation is done into DefaultCacheImpl and not exposed
to DefaultCache public API. New APIs will be exposed in a separate
commit when Size() for the protected cache will be implemented.

Relates-To: OLPEDGE-2392

Signed-off-by: Mykola Malik <ext-mykola.malik@here.com>