Reduce extra quads count for prefetch.
In case of sparse data we load additional quads, but current algorithm
looks on slice root instead of requested tiles when decides to load
extra metadata. In some cases that is not correct because for all
requested tiles bundles are not that far away.

Volatile client doesn't load extra quads, thus, only versioned client
changed.

Also added the test to cover the issue.

Resolves: OLPEDGE-2454

Signed-off-by: Kostiantyn Zvieriev <ext-kostiantyn.zvieriev@here.com>