Add SignInApple to Authentication API
This is the next step in the Apple Sign-In support.
Using 'AppleSignInProperties' user could sign-in in the HERE services.
Apple id_token will be used as a bearer. For acceptance T&C user still
required to have app_id and app_secret.

Work on this API is still in progress thus it may be changed in the
future

Relates-To: OLPEDGE-2483
Signed-off-by: Andrey Kashcheev <ext-andrey.kashcheev@here.com>