Use the server time correctly
Fix the various APIs in authentication component to use the server
time when the `use_system_time` option is false.

Server time is now incremented on retry.

Now we always try to retry with a server time on a wrong timestamp
error.

Add a bool operator to ApiResponse class, a shortcut to IsSuccessful.

Relates-To: OLPSUP-14833

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>