Wrap the leveldb::Env to shared_ptr
Change the static EnvWrapper instance to a wrapped with a smart
pointer, so that we can guaranty that the instance will be alive while
it is needed, even when the `env` was deinitialized at exit.
By this we fix the scenario when the environment is destroyed before
DiskCache in certain use cases.

Resolves: OLPSUP-14933

Signed-off-by: Mykhailo Kuchma <ext-mykhailo.kuchma@here.com>