Extend TokenProvider with cancellable operator()
This operator returns Response<TokenResult>,
so that a caller can get error details in case of
retrieving token failure.

This is the initial implementation. The usage of
CancellationContext is not implemented yet,
so the token request cannot be cancelled for now.

Old TokenProvider::operator() returning a string
is marked as deprecated.

TokenResult methods GetHttpStatus() and
GetErrorResponse() are marked as deprecated.

SignInResult and BaseResult classes are extended
with a method GetFullMessage(). This method
returns string representation of a network
response json if it contain an error.

Changed a logic of error propagation inside
TokenEndpoint::Impl::RequestToken() method. Now,
if SignInResult contains an error, the error data
is passed to the callback as ApiError and not
TokenResult containing an error, as it used to be
before this commit. This approach gives the user
an ability to check if token is valid in sindle
call.

Relates-To: OLPEDGE-2598

Signed-off-by: Mykola Malik <ext-mykola.malik@here.com>