Add LMDB to external. Add CMake for LMDB.
lmdb is added to external libraries. CMakeLists.txt.lmdb.in and
lmdbConfig.cmake.in was created for lmdb.
Changes in olp-cpp-sdk-core and external CMakeLists.txt was made.
Added OLP_SDK_ENABLE_DEFAULT_CACHE_LMDB option to be able to
enable/disable build of lmdb. README.md was changed with
description of a new flag.

Resolves: OLPEDGE-2645

Signed-off-by: Yevhen Krasilnyk <ext-yevhen.krasilnyk@here.com>