Add proper cancellation for the new token provider
The inner class TokenEndpoint::Impl is extructed from TokenEndpoint.cpp
to separate files TokenEndpointImpl.h and TokenEndpointImpl.cpp.

TokenEndpointImpl is extended with mostly copy-pasted code from
AuthenticationClientImpl, which is supposed to be refactored in
the future to use TokenProvider, and not in the opposite direction,
as it is now. TokenEndpointImpl code is adapted in a way to give an
abilty to pass cancellation context from the new token profider to
OlpClient::CallApi().

A new test class TokenEndpointTest contains adapted tests
copy-pasted from AuthenticationClientTest.

Integration tests for cancellation and timeout are added.

Relates-To: OLPEDGE-2651

Signed-off-by: Mykola Malik <ext-mykola.malik@here.com>