Make sign in timeout error distinguishable
If a timeout occurred, the cancellation is done through the context.
So this case needs to be handled independently of context state,
instead of blindely return Cancelled error if context is cancelled.

Integration tests for timeout are to AuthenticationClientTest class.
Some other minor improvements performed for AuthenticationClientTest,
like fixing typos and decreasing code duplications.

Relates-To: OLPEDGE-2656

Signed-off-by: Mykola Malik <ext-mykola.malik@here.com>