Increase libcurl version needed and detected.
Pull request #1317 introduced the HTTPS proxy support for cURL based
Linux network layer which required a libcurl higher or equal to version
7.52.0, see https://curl.se/libcurl/c/CURLOPT_PROXYTYPE.html.
To make sure we do not break customers which have manually installed
libcurl the cmake find_package needed to be adapted so that the user
gets a early error during cmake configuration.
Same applies for the README.md file so that the user has the
information which he needs without reading the code.

Relates-To: OLPEDGE-1099

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>