Add cache optimizations
Replace substr to memcmp to avoid additional allocations
and put it to the IsPrefix function.
Put checking memory cache from the first cache in the order
for checking Contains to the last one as it usually small
in size and and not sufficient to store the entire cache.

Add checks expiry validity in case of LRU cache as there is no
need for additional checks if the expiry is not valid
Check protected_data size in IsProtected if there are elements
in ProtectedKeyList.

Relates-To: OLPSUP-18424

Signed-off-by: Yevhenii Dudnyk <ext-yevhenii.dudnyk@here.com>