Separate thread for cancellation tasks
This commit introduces separate thread specifically for the cancellation
tasks. By doing so we may speed up final callback execution as task
won't wait for its turn in the main busy queue and instead will be
executed on the separate usually free thread.
This is a quite bold move as we always allocate one additional thread
regardless of user-specified total threads number.

Relates-To: OLPSUP-17825
Signed-off-by: Andrey Kashcheev <ext-andrey.kashcheev@here.com>