Wrap PrefetchTiles call into CancellationContext. (#1336)
This helps speed-up cancelling of the operation as chain of contexts
doesn't break. Previously to this, the `execution_context` was not used
internally in the `PrefetchTilesHelper::Prefetch` method which
resulted in a break of the cancellation context chain and thus
it was not able to cancel all subsequently triggered prefetch
jobs from the one master context, which was handed over to the caller
of the VersionedLayerClient::PrefetchTiles() API. Due to this then
when user actually cancelled the request it would still have to wait for
the download to finish until the callback came. Now this is fixed.

Relates-To: OLPSUP-17825
Signed-off-by: Andrey Kashcheev <ext-andrey.kashcheev@here.com>
Signed-off-by: Andrei Popescu <andrei.popescu@here.com>