Fix olp::utils::Dir::Size() recursive behaviour.
This commit fixes the recursive behaviour of the olp::utils::Dir::Size()
method to go all the way deep and find all not filtered out files to
consider for calculating the dir size.
This is done by using a queue to push every found dir inside the
currently scanned dir onto it for further dept scanning.

Relates-To: OLPEDGE-2700

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>