Open cache in r/o mode on `no space left` errors
Since protected cache could be used as a static cache it is expected
to be able to use it even when there is no available space on device.
But when protected cache is being open in R/W mode levelDb may try to
create some files and this failure would be treated later as a corrupted
cache.
Instead, on `no space left` errors we will try to open the cache again
but in read-only mode to enforce static behavior.

Relates-To: OLPSUP-21193
Signed-off-by: Andrey Kashcheev <ext-andrey.kashcheev@here.com>